+++ components/Tabs/Tab.less
@@ -1,6 +1,25 @@
+@import '../variables.less';
+
+@easing: 0.2s ease-out;
+
 :local {
   .root {
-    // padding: 0 40px;
-    // margin: 40px 0;
+    display: inline-block;
+    margin-left: 20px;
+    margin-right: 20px;
+    box-sizing: border-box;
+    line-height: 41px;
+    font-size: 16px;
+    border-bottom: 3px solid transparent;
+    transition: border-bottom @easing;
+    cursor: pointer;
+  }
+
+  .root:hover {
+    border-bottom: 3px solid @border-color-focus-light;
+  }
+
+  .active {
+    cursor: default;
   }
 }
