+++ components/Tabs/Tabs.less
@@ -1,5 +1,7 @@
 :local {
   .root {
-    display: block;
+    margin: 0;
+    padding: 0;
+    position: relative;
   }
 }
