--- components/Tabs/index.js
@@ -1,3 +1,0 @@
-// @flow
-import Tabs from './Tabs';
-module.exports = Tabs;
