--- components/TabsV2/Tab.less
@@ -1,25 +1,0 @@
-@import '../variables.less';
-
-@easing: 0.2s ease-out;
-
-:local {
-  .root {
-    display: inline-block;
-    margin-left: 20px;
-    margin-right: 20px;
-    box-sizing: border-box;
-    line-height: 41px;
-    font-size: 16px;
-    border-bottom: 3px solid transparent;
-    transition: border-bottom @easing;
-    cursor: pointer;
-  }
-
-  .root:hover {
-    border-bottom: 3px solid @border-color-focus-light;
-  }
-
-  .active {
-    cursor: default;
-  }
-}
