--- components/TabsV2/Tabs.less
@@ -1,7 +1,0 @@
-:local {
-  .root {
-    margin: 0;
-    padding: 0;
-    position: relative;
-  }
-}
